+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Tools development version
 
 - Allow relaxed version with only major and minor components in `match_semver()` with `strict_semver=False`. (#49, @kelly-sovacool)
+- Remove `args` and add `repo` parameter to `get_latest_release_tag()` and `get_latest_release_hash()`. (#51, @kelly-sovacool)
 
 ## Tools 0.2.4